minio_microservice/requirements.txt
minio
flask
python-dotenv
